requirements.txt
fastapi
uvicorn
cryptography
motor
passlib
hmac
datetime
contextlib
base64
pydantic
python-jose
cryptography
hashlib
os
pymongo
uuid